rag-agent-chat/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pypdf
python-dotenv
requests
python-multipart
pydantic
